Extract mockHttpClient for testing
